[SecurityBundle] decouple the logout PHP helper and Twig extension
