+++ src/Symfony/Bundle/SecurityBundle/Resources/config/templating_php.xml
@@ -12,9 +12,7 @@
     <services>
         <service id="templating.helper.logout_url" class="%templating.helper.logout_url.class%">
             <tag name="templating.helper" alias="logout_url" />
-            <argument type="service" id="request_stack" />
-            <argument type="service" id="router" />
-            <argument type="service" id="security.token_storage" />
+            <argument type="service" id="security.logout_url_generator" />
         </service>
 
         <service id="templating.helper.security" class="%templating.helper.security.class%">
