+++ src/Symfony/Bundle/SecurityBundle/Resources/config/templating_twig.xml
@@ -5,14 +5,14 @@
     xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
 
     <parameters>
-        <parameter key="twig.extension.logout_url.class">Symfony\Bundle\SecurityBundle\Twig\Extension\LogoutUrlExtension</parameter>
+        <parameter key="twig.extension.logout_url.class">Symfony\Bridge\Twig\Extension\LogoutUrlExtension</parameter>
         <parameter key="twig.extension.security.class">Symfony\Bridge\Twig\Extension\SecurityExtension</parameter>
     </parameters>
 
     <services>
         <service id="twig.extension.logout_url" class="%twig.extension.logout_url.class%" public="false">
             <tag name="twig.extension" />
-            <argument type="service" id="templating.helper.logout_url" />
+            <argument type="service" id="security.logout_url_generator" />
         </service>
 
         <service id="twig.extension.security" class="%twig.extension.security.class%" public="false">
